Split out gen_maton_f out of photo_gen_matom.c
